+++ tests/integration/long_callback/app_side_update.py
@@ -17,20 +17,20 @@ app.layout = html.Div(
         html.Div(id="side-status"),
     ]
 )
+app.test_lock = lock = long_callback_manager.test_lock
 
 
 @callback(
     Output("result", "children"),
     [Input("run-button", "n_clicks")],
-    long=True,
-    long_progress=Output("status", "children"),
-    long_progress_default="Finished",
-    long_cancel=[Input("cancel-button", "n_clicks")],
-    long_interval=0,
+    background=True,
+    progress=Output("status", "children"),
+    progress_default="Finished",
+    cancel=[Input("cancel-button", "n_clicks")],
+    interval=0,
     prevent_initial_call=True,
 )
 def update_output(set_progress, n_clicks):
-    print("trigger")
     for i in range(4):
         set_progress(f"Progress {i}/4")
         time.sleep(1)